refactor: use customViews with custom attributes
