+++ app/src/main/java/com/alejandraacevedo/mymovies/customviews/AspectRatioImageView.kt
@@ -3,11 +3,18 @@ package com.alejandraacevedo.mymovies.customviews
 import android.content.Context
 import android.util.AttributeSet
 import androidx.appcompat.widget.AppCompatImageView
+import com.alejandraacevedo.mymovies.R
 
 class AspectRatioImageView @JvmOverloads constructor(
         context: Context, attrs: AttributeSet? = null, defStyleAttr: Int = 0
 ) : AppCompatImageView(context, attrs, defStyleAttr) {
-    var ratio: Float = 1f
+    private var ratio: Float = 1f
+
+    init {
+        val a = context.obtainStyledAttributes(attrs, R.styleable.AspectRatioImageView)
+        ratio = a.getFloat(R.styleable.AspectRatioImageView_ratio, 1f)
+        a.recycle()
+    }
 
     override fun onMeasure(widthMeasureSpec: Int, heightMeasureSpec: Int) {
         super.onMeasure(widthMeasureSpec, heightMeasureSpec)
