+++ app/src/main/res/layout/activity_detail.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".DetailActivity">
+    tools:context=".activities.DetailActivity">
 
     <ImageView
         android:id="@+id/backdrop"
