+++ app/src/main/res/values/attrs.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <declare-styleable name="AspectRatioImageView">
+        <attr name="ratio" format="float"></attr>
+    </declare-styleable></resources>
